:recycle: api: Move logger to src
